+++ include/robot_interfaces_solo/solo12_driver.hpp
@@ -1,6 +1,6 @@
 /**
  * \file
- * \brief Driver for using Solo12 in robot_interfaces::RobotBackend.
+ * \brief Driver for using BoltHumanoid in robot_interfaces::RobotBackend.
  * \copyright Copyright (c) 2022, Max Planck Gesellschaft.
  */
 #pragma once
@@ -19,13 +19,13 @@
 
 namespace robot_interfaces_solo
 {
-//! Driver to use Solo12
-class Solo12Driver : public BaseSolo12Driver
+//! Driver to use BoltHumanoid
+class BoltHumanoidDriver : public BaseBoltHumanoidDriver
 {
 public:
-    inline static const std::string LOGGER_NAME = "Solo12Driver";
+    inline static const std::string LOGGER_NAME = "BoltHumanoidDriver";
 
-    explicit Solo12Driver(const Solo12Config &config);
+    explicit BoltHumanoidDriver(const BoltHumanoidConfig &config);
 
     // RobotDriver methods
     void initialize() override;
@@ -36,20 +36,20 @@ public:
 
 private:
     std::shared_ptr<spdlog::logger> log_;
-    const Solo12Config config_;
-    solo::Solo12 solo12_;
+    const BoltHumanoidConfig config_;
+    solo::BoltHumanoid solo12_;
     Action applied_action_;
     bool is_initialized_ = false;
 };
 
 //! Fake driver for testing (ignores actions and returns some artificial
 //! observations).
-class FakeSolo12Driver : public BaseSolo12Driver
+class FakeBoltHumanoidDriver : public BaseBoltHumanoidDriver
 {
 public:
-    inline static const std::string LOGGER_NAME = "FakeSolo12Driver";
+    inline static const std::string LOGGER_NAME = "FakeBoltHumanoidDriver";
 
-    explicit FakeSolo12Driver(const Solo12Config &config);
+    explicit FakeBoltHumanoidDriver(const BoltHumanoidConfig &config);
 
     // RobotDriver methods
     void initialize() override;
@@ -60,16 +60,16 @@ public:
 
 private:
     std::shared_ptr<spdlog::logger> log_;
-    const Solo12Config config_;
+    const BoltHumanoidConfig config_;
     Action applied_action_;
     bool is_initialized_ = false;
     double t_ = 0.0;
 };
 
 /**
- * @brief Create robot backend using the Solo12 driver (real robot).
+ * @brief Create robot backend using the BoltHumanoid driver (real robot).
  *
- * @see Solo12Driver
+ * @see BoltHumanoidDriver
  *
  * @param robot_data  Instance of RobotData used for communication.
  * @param driver_config  Driver configuration.
@@ -78,25 +78,25 @@ private:
  * @param max_number_of_actions  Number of actions after which the backend
  *     automatically shuts down.
  *
- * @return A RobotBackend instances which uses a Solo12 driver.
+ * @return A RobotBackend instances which uses a BoltHumanoid driver.
  */
-Solo12Backend::Ptr create_real_solo12_backend(
-    Solo12Data::Ptr robot_data,
-    const Solo12Config &driver_config,
+BoltHumanoidBackend::Ptr create_real_solo12_backend(
+    BoltHumanoidData::Ptr robot_data,
+    const BoltHumanoidConfig &driver_config,
     const double first_action_timeout = std::numeric_limits<double>::infinity(),
     const uint32_t max_number_of_actions = 0);
 
 /**
- * @brief Create robot backend using the fake Solo12 driver (for testing).
+ * @brief Create robot backend using the fake BoltHumanoid driver (for testing).
  *
  * Arguments are the same as for @ref create_real_solo12_backend.
  *
- * @see FakeSolo12Driver
+ * @see FakeBoltHumanoidDriver
  * @see create_real_solo12_backend
  */
-Solo12Backend::Ptr create_fake_solo12_backend(
-    Solo12Data::Ptr robot_data,
-    const Solo12Config &driver_config,
+BoltHumanoidBackend::Ptr create_fake_solo12_backend(
+    BoltHumanoidData::Ptr robot_data,
+    const BoltHumanoidConfig &driver_config,
     const double first_action_timeout = std::numeric_limits<double>::infinity(),
     const uint32_t max_number_of_actions = 0);