+++ include/robot_interfaces_solo/solo12_utils.hpp
@@ -1,6 +1,6 @@
 /**
  * \file
- * \brief Utility functions for Solo12.
+ * \brief Utility functions for BoltHumanoid.
  * \copyright Copyright (c) 2022, Max Planck Gesellschaft.
  */
 #pragma once
@@ -12,7 +12,7 @@
 namespace robot_interfaces_solo
 {
 /**
- * @brief Create robot backend using a Solo12 driver.
+ * @brief Create robot backend using a BoltHumanoid driver.
  *
  * @param robot_data  Instance of RobotData used for communication.
  * @param robot_driver  Driver instance, connecting to the robot or simulation.
@@ -25,11 +25,11 @@ namespace robot_interfaces_solo
  *     This should enabled if using the real robot but may be disabled when
  *     using simulation.
  *
- * @return A RobotBackend instances which uses a Solo12 driver.
+ * @return A RobotBackend instances which uses a BoltHumanoid driver.
  */
-Solo12Backend::Ptr create_solo12_backend(
-    Solo12Data::Ptr robot_data,
-    BaseSolo12Driver::Ptr robot_driver,
+BoltHumanoidBackend::Ptr create_solo12_backend(
+    BoltHumanoidData::Ptr robot_data,
+    BaseBoltHumanoidDriver::Ptr robot_driver,
     const double first_action_timeout = std::numeric_limits<double>::infinity(),
     const uint32_t max_number_of_actions = 0,
     bool enable_timing_watchdog = true);