+++ src/solo12_driver.cpp
@@ -14,7 +14,7 @@
 
 namespace robot_interfaces_solo
 {
-Solo12Driver::Solo12Driver(const Solo12Config &config) : config_(config)
+BoltHumanoidDriver::BoltHumanoidDriver(const BoltHumanoidConfig &config) : config_(config)
 {
     // initialise logger and set level based on config
     log_ = spdlog::get(LOGGER_NAME);
@@ -26,9 +26,9 @@ Solo12Driver::Solo12Driver(const Solo12Config &config) : config_(config)
     }
 }
 
-void Solo12Driver::initialize()
+void BoltHumanoidDriver::initialize()
 {
-    log_->debug("Initialize Solo12");
+    log_->debug("Initialize BoltHumanoid");
     solo12_.initialize(config_.network_interface, config_.slider_serial_port);
     solo12_.set_max_current(config_.max_motor_current_A);
 
@@ -64,7 +64,7 @@ void Solo12Driver::initialize()
     log_->debug("Initialization finished");
 }
 
-Solo12Driver::Action Solo12Driver::apply_action(const Action &desired_action)
+BoltHumanoidDriver::Action BoltHumanoidDriver::apply_action(const Action &desired_action)
 {
     double start_time_sec = real_time_tools::Timer::get_current_time_sec();
 
@@ -94,7 +94,7 @@ Solo12Driver::Action Solo12Driver::apply_action(const Action &desired_action)
     return applied_action_;
 }
 
-Solo12Driver::Observation Solo12Driver::get_latest_observation()
+BoltHumanoidDriver::Observation BoltHumanoidDriver::get_latest_observation()
 {
     if (!is_initialized_)
     {
@@ -127,7 +127,7 @@ Solo12Driver::Observation Solo12Driver::get_latest_observation()
     return obs;
 }
 
-std::string Solo12Driver::get_error()
+std::string BoltHumanoidDriver::get_error()
 {
     std::string error_msg = "";
 
@@ -164,7 +164,7 @@ std::string Solo12Driver::get_error()
     return error_msg;
 }
 
-void Solo12Driver::shutdown()
+void BoltHumanoidDriver::shutdown()
 {
     // TODO: is there a way to completely disable motors?
     if (is_initialized_)
@@ -173,7 +173,7 @@ void Solo12Driver::shutdown()
     }
 }
 
-FakeSolo12Driver::FakeSolo12Driver(const Solo12Config &config) : config_(config)
+FakeBoltHumanoidDriver::FakeBoltHumanoidDriver(const BoltHumanoidConfig &config) : config_(config)
 {
     // initialise logger and set level based on config
     log_ = spdlog::get(LOGGER_NAME);
@@ -185,14 +185,14 @@ FakeSolo12Driver::FakeSolo12Driver(const Solo12Config &config) : config_(config)
     }
 }
 
-void FakeSolo12Driver::initialize()
+void FakeBoltHumanoidDriver::initialize()
 {
-    log_->debug("Initialize Fake Solo12");
+    log_->debug("Initialize Fake BoltHumanoid");
     real_time_tools::Timer::sleep_sec(2);
     is_initialized_ = true;
 }
 
-FakeSolo12Driver::Action FakeSolo12Driver::apply_action(
+FakeBoltHumanoidDriver::Action FakeBoltHumanoidDriver::apply_action(
     const Action &desired_action)
 {
     double start_time_sec = real_time_tools::Timer::get_current_time_sec();
@@ -209,7 +209,7 @@ FakeSolo12Driver::Action FakeSolo12Driver::apply_action(
     return desired_action;
 }
 
-FakeSolo12Driver::Observation FakeSolo12Driver::get_latest_observation()
+FakeBoltHumanoidDriver::Observation FakeBoltHumanoidDriver::get_latest_observation()
 {
     if (!is_initialized_)
     {
@@ -247,36 +247,36 @@ FakeSolo12Driver::Observation FakeSolo12Driver::get_latest_observation()
     return obs;
 }
 
-std::string FakeSolo12Driver::get_error()
+std::string FakeBoltHumanoidDriver::get_error()
 {
     return "";
 }
 
-void FakeSolo12Driver::shutdown()
+void FakeBoltHumanoidDriver::shutdown()
 {
 }
 
-Solo12Backend::Ptr create_real_solo12_backend(
-    Solo12Data::Ptr robot_data,
-    const Solo12Config &driver_config,
+BoltHumanoidBackend::Ptr create_real_solo12_backend(
+    BoltHumanoidData::Ptr robot_data,
+    const BoltHumanoidConfig &driver_config,
     const double first_action_timeout,
     const uint32_t max_number_of_actions)
 {
     constexpr bool enable_timing_watchdog = true;
     return create_solo12_backend(robot_data,
-                                 std::make_shared<Solo12Driver>(driver_config),
+                                 std::make_shared<BoltHumanoidDriver>(driver_config),
                                  first_action_timeout,
                                  max_number_of_actions,
                                  enable_timing_watchdog);
 }
 
-Solo12Backend::Ptr create_fake_solo12_backend(
-    Solo12Data::Ptr robot_data,
-    const Solo12Config &driver_config,
+BoltHumanoidBackend::Ptr create_fake_solo12_backend(
+    BoltHumanoidData::Ptr robot_data,
+    const BoltHumanoidConfig &driver_config,
     const double first_action_timeout,
     const uint32_t max_number_of_actions)
 {
-    auto driver = std::make_shared<FakeSolo12Driver>(driver_config);
+    auto driver = std::make_shared<FakeBoltHumanoidDriver>(driver_config);
 
     constexpr bool enable_timing_watchdog = false;
     return create_solo12_backend(robot_data,